+++ letterFrequencyInMapping/src/main/java/it/unipi/hadoop/LetterFrequency.java
@@ -6,7 +6,7 @@ import java.util.Map;
 import org.apache.hadoop.io.Text;
 import org.apache.hadoop.conf.Configuration;
 import org.apache.hadoop.fs.Path;
-import org.apache.hadoop.io.IntWritable;
+import org.apache.hadoop.io.LongWritable;
 import org.apache.hadoop.io.DoubleWritable;
 import org.apache.hadoop.mapreduce.Job;
 import org.apache.hadoop.mapreduce.Mapper;
@@ -21,16 +21,16 @@ import java.util.regex.Pattern;
 public class LetterFrequency
 {
 
-    public static class FrequencyMapper extends Mapper<Object, Text, Text, IntWritable> 
+    public static class FrequencyMapper extends Mapper<Object, Text, Text, LongWritable> 
     {
-        private Map<Text, IntWritable> map;
+        private Map<Text, LongWritable> map;
         private final static Pattern CHARACTER_PATTERN = Pattern.compile("[a-z]", Pattern.CASE_INSENSITIVE);
 
         @Override
         public void setup(Context context)
         {
             // Initialize the map
-            map = new HashMap<Text, IntWritable>();
+            map = new HashMap<Text, LongWritable>();
         }
 
         @Override
@@ -44,9 +44,9 @@ public class LetterFrequency
                 // Check if the character is a letter
                 if (CHARACTER_PATTERN.matcher(String.valueOf(ch)).matches()) { 
                     if (map.containsKey(new Text(String.valueOf(ch)))) {
-                        map.put(new Text(String.valueOf(ch)), new IntWritable(map.get(new Text(String.valueOf(ch))).get() + 1));
+                        map.put(new Text(String.valueOf(ch)), new LongWritable(map.get(new Text(String.valueOf(ch))).get() + 1));
                     } else {
-                        map.put(new Text(String.valueOf(ch)), new IntWritable(1));
+                        map.put(new Text(String.valueOf(ch)), new LongWritable(1));
                     }
                 }
             }
@@ -62,26 +62,26 @@ public class LetterFrequency
         }
     }
 
-    public static class FrequencyReducer extends Reducer<Text, IntWritable, Text, DoubleWritable>
+    public static class FrequencyReducer extends Reducer<Text, LongWritable, Text, DoubleWritable>
     {
         // Variables
-        private static int TEXT_LENGTH;
+        private static long TEXT_LENGTH;
 
         public void setup(Context context)
         {
             // Configuration
             Configuration conf = context.getConfiguration();
-            TEXT_LENGTH = Integer.parseInt(conf.get("textLength"));
+            TEXT_LENGTH = Long.parseLong(conf.get("textLength"));
         }
 
         @Override
-        public void reduce(Text key, Iterable<IntWritable> values, Context context) throws IOException, InterruptedException 
+        public void reduce(Text key, Iterable<LongWritable> values, Context context) throws IOException, InterruptedException 
         {
             // Variables
             int sum = 0;
 
             // Iterate over the values
-            for (IntWritable value : values) {
+            for (LongWritable value : values) {
                 sum += value.get();
             }
 
@@ -91,11 +91,11 @@ public class LetterFrequency
         }
     }
 
-    public static Job configureJob(Configuration conf, Map<String, String> argMap, int textLength, int numReducerTasks) throws IOException {
+    public static Job configureJob(Configuration conf, Map<String, String> argMap, long textLength, int numReducerTasks) throws IOException {
         Job letterFrequencyJob = Job.getInstance(conf, "LetterFrequency");
     
         // Set the configuration
-        letterFrequencyJob.getConfiguration().setInt("textLength", textLength);
+        letterFrequencyJob.getConfiguration().setLong("textLength", textLength);
     
         // Set the main classes
         letterFrequencyJob.setJarByClass(LetterFrequency.class);
@@ -111,7 +111,7 @@ public class LetterFrequency
     
         // Set the output key and value classes for the mapper
         letterFrequencyJob.setMapOutputKeyClass(Text.class);
-        letterFrequencyJob.setMapOutputValueClass(IntWritable.class);
+        letterFrequencyJob.setMapOutputValueClass(LongWritable.class);
     
         // Set the output key and value classes for the reducer
         letterFrequencyJob.setOutputKeyClass(Text.class);
